+++ packages/evokit-box/__tests__/Box.height.test.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+import barista from 'seed-barista';
+import style from '../style.css';
 import { Box } from '../src';
 
 describe('<Box box-height="..." />', () => {
@@ -8,5 +10,12 @@ describe('<Box box-height="..." />', () => {
         expect(wrapper).toMatchSnapshot();
 
         expect(wrapper.html()).toEqual('<div class="ek-box ek-box_height_1-1"></div>');
+    });
+
+    it('css', () => {
+        const css = barista({ content: style });
+
+        expect(css.rule('.ek-box_height_auto').exists()).toBeTruthy();
+        expect(css.rule('.ek-box_height_1-1').exists()).toBeTruthy();
     });
 });